+++ action_server/tests/action_server_tests/test_server_secrets/pack_encryption/hello_action.py
@@ -6,6 +6,9 @@ def get_private(private_info: Secret) -> str:
     """
     Returns the value passed to the private key.
 
+    Args:
+        private_info: Some private info.
+
     Returns:
         The value of the private key.
     """